refactor: split civilian units panel support into scoped parts (#3878)
Decompose the 676-line civilian_units_panel_support.dart part into
resolution, row-card chrome, and unit-row builders so each concern stays
under the panel submodule size target without behavior changes.

Refs #3878
