+++ app/lib/features/game/widgets/civilian_units_panel_support_unit_row.dart
@@ -1,110 +1,6 @@
-/// Human-readable label for work target ids. SPEC/ui/civilian-units-panel.md.
+/// Per-unit civilian row builder. SPEC/ui/civilian-units-panel.md.
 
 part of 'civilian_units_panel.dart';
-
-const Map<String, String> _workTargetLabels = {
-  kWorkTargetExplore: 'Explore',
-  kWorkTargetProspect: 'Prospect',
-  kWorkTargetBuildImprovement: 'Build improvement',
-  kWorkTargetUpgradeTown: 'Upgrade town',
-  kWorkTargetBuildRoad: 'Build road',
-  kWorkTargetBuildPort: 'Build port',
-  kWorkTargetBuildFort: 'Build fort',
-  kWorkTargetBuildRail: 'Build rail',
-  kWorkTargetCounterSpy: 'Counter spy',
-  kWorkTargetPurchaseLand: 'Purchase land',
-};
-
-// Sort/partition helpers live in `civilian_units_sort.dart` (public surface):
-// `provinceNamesByPrefixedId`, `isCivilianUnit`, `civilianUnitsInRegion`, and
-// `civilianSortProvinceName`. They are imported by the panel library root
-// (`civilian_units_panel.dart`) and visible here through the shared library
-// scope. Refs #2575 (Phase 4 testability).
-
-/// Pending assigned-to line plus optional cost strip. SPEC/ui/civilian-units-panel.md.
-class _PendingAssignedResolution {
-  const _PendingAssignedResolution({
-    required this.mainLine,
-    required this.totalTurns,
-    this.materialCosts,
-    this.treasuryAmount,
-  });
-
-  final String mainLine;
-  final int totalTurns;
-  final Map<String, int>? materialCosts;
-  final int? treasuryAmount;
-}
-
-_PendingAssignedResolution _resolvePendingAssignedResolution(
-  Game game,
-  Unit unit,
-  WorkOrder order,
-  Map<String, String> provinceNames,
-) {
-  final workLabel = _workTargetLabels[order.target] ?? order.target;
-  final regionId = Unit.regionIdFromTileKey(order.targetTileKey);
-  final provinceId = Unit.provinceIdFromTileKey(order.targetTileKey);
-  var location = '';
-  if (regionId != null && provinceId != null) {
-    final name =
-        provinceNames['$regionId|$provinceId'] ?? '$regionId|$provinceId';
-    location = ' (${regionDisplayLabel(regionId)} — $name)';
-  }
-  final base = '$workLabel$location';
-  final totalTurns = previewTotalTurnsForPendingWorkOrder(
-    game: game,
-    unit: unit,
-    order: order,
-  );
-
-  if (order.target == kWorkTargetPurchaseLand) {
-    final resourceId = game.worldState.resourceByTileKey[order.targetTileKey];
-    if (resourceId != null && resourceId.isNotEmpty) {
-      return _PendingAssignedResolution(
-        mainLine: base,
-        totalTurns: totalTurns,
-        treasuryAmount: purchaseLandCost(resourceId),
-      );
-    }
-    return _PendingAssignedResolution(mainLine: base, totalTurns: totalTurns);
-  }
-  if (order.target == kWorkTargetCounterSpy) {
-    return _PendingAssignedResolution(mainLine: base, totalTurns: totalTurns);
-  }
-
-  final targetProvinceId = Unit.provinceIdFromTileKey(order.targetTileKey);
-  final province = targetProvinceId != null
-      ? game.worldState.tryGetProvince(targetProvinceId)
-      : null;
-
-  final improvementLevel = order.target == kWorkTargetBuildImprovement
-      ? game.worldState.tileState.improvementLevel(order.targetTileKey)
-      : 0;
-  final fortLevel = province?.fortLevel ?? 0;
-  final roadLevel = game.worldState.tileState.roadLevel(order.targetTileKey);
-
-  final costMap = WorkOrderCostCalculator(game).calculateCost(
-    order.target,
-    order.targetTileKey,
-    improvementLevel: improvementLevel,
-    fortLevel: fortLevel,
-    roadLevel: roadLevel,
-  );
-  if (costMap != null && costMap.isNotEmpty) {
-    return _PendingAssignedResolution(
-      mainLine: base,
-      totalTurns: totalTurns,
-      materialCosts: costMap,
-    );
-  }
-  return _PendingAssignedResolution(mainLine: base, totalTurns: totalTurns);
-}
-
-List<MapEntry<String, int>> _sortedMaterialCostEntries(Map<String, int> m) {
-  final list = m.entries.toList()..sort((a, b) => a.key.compareTo(b.key));
-  return list;
-}
 
 class _UnitRow extends ConsumerWidget {
   const _UnitRow({
@@ -498,179 +394,5 @@ class _UnitRow extends ConsumerWidget {
       ),
       actions: rowActions,
     );
-  }
-}
-
-/// Bordered civilian-unit row card per `SPEC/ui/civilian-units-panel.md`
-/// § Layout / wireframe → Row card chrome and mockup `.unit-row`
-/// (Refs #2866 S9 R30).
-///
-/// Wraps the left detail stack and the right-aligned action cluster inside a
-/// single `DecoratedBox` painted with a vertical
-/// `EditorialMonoclePalette.bgDeep` → `EditorialMonoclePalette.surface`
-/// gradient and a 1 dp border. The default border resolves to
-/// [EditorialMonoclePalette.border]; pointer hover (via [MouseRegion]) and
-/// tile-scope [selected] state both shift the border to
-/// [EditorialMonoclePalette.accentDim] so the card surfaces selection /
-/// pointer feedback without relying on Material `ListTile` chrome.
-///
-/// Public so widget tests can locate civilian rows via
-/// `find.byType(CivilianUnitRowCard)` after the migration off `ListTile`.
-class CivilianUnitRowCard extends StatefulWidget {
-  const CivilianUnitRowCard({
-    super.key,
-    required this.details,
-    required this.actions,
-    required this.selected,
-    required this.onTap,
-  });
-
-  final Widget details;
-  final List<UnitsEntityAction> actions;
-  final bool selected;
-  final VoidCallback onTap;
-
-  @override
-  State<CivilianUnitRowCard> createState() => _CivilianUnitRowCardState();
-}
-
-class _CivilianUnitRowCardState extends State<CivilianUnitRowCard> {
-  bool _hovered = false;
-
-  static const double _borderWidth = 1;
-  static const double _innerSpacing = 8;
-  static const double _minRowHeight = 44;
-
-  Color _borderColor() {
-    if (widget.selected || _hovered) {
-      return EditorialMonoclePalette.accentDim;
-    }
-    return EditorialMonoclePalette.border;
-  }
-
-  static final LinearGradient _cardGradient = LinearGradient(
-    begin: Alignment.topCenter,
-    end: Alignment.bottomCenter,
-    colors: <Color>[
-      EditorialMonoclePalette.bgDeep,
-      EditorialMonoclePalette.surface,
-    ],
-  );
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 4),
-      child: MouseRegion(
-        onEnter: (_) {
-          if (!_hovered) setState(() => _hovered = true);
-        },
-        onExit: (_) {
-          if (_hovered) setState(() => _hovered = false);
-        },
-        child: Material(
-          type: MaterialType.transparency,
-          child: InkWell(
-            onTap: widget.onTap,
-            child: DecoratedBox(
-              decoration: BoxDecoration(
-                gradient: _cardGradient,
-                border: Border.all(color: _borderColor(), width: _borderWidth),
-              ),
-              child: Padding(
-                padding: const EdgeInsets.all(_innerSpacing),
-                child: ConstrainedBox(
-                  constraints: const BoxConstraints(minHeight: _minRowHeight),
-                  child: Row(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Expanded(child: widget.details),
-                      if (widget.actions.isNotEmpty) ...[
-                        CtGap.wm,
-                        _CivilianUnitCardActions(actions: widget.actions),
-                      ],
-                    ],
-                  ),
-                ),
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-/// Right-aligned action cluster inside [CivilianUnitRowCard]. Renders the
-/// mockup compact-pill row actions per `SPEC/ui/civilian-units-panel.md`
-/// § Row actions and the `UNIT10001` mockup `.u-actions` family (issue #3514
-/// owner decisions #6/#7):
-///
-/// - neutral actions (e.g. **Assign**) render as [CtActionTextButton] pills
-///   with an icon + label (mockup `.u-actions button`),
-/// - destructive actions ([UnitsEntityActionVariant.danger], e.g. **Cancel**)
-///   render as [CtDangerTextButton] pills (mockup `.u-actions .cancel-btn`),
-/// - `iconOnly` actions (the rightmost **Locate** control per R30) render as a
-///   circular [CtCircularLocateButton] (mockup `.u-actions .locate-btn`).
-///
-/// The cluster stays a right-aligned [Wrap] so it flows onto a second line at
-/// narrow widths rather than overflowing horizontally.
-class _CivilianUnitCardActions extends StatelessWidget {
-  const _CivilianUnitCardActions({required this.actions});
-
-  final List<UnitsEntityAction> actions;
-
-  static const double _spacing = 6;
-
-  @override
-  Widget build(BuildContext context) {
-    return Wrap(
-      spacing: _spacing,
-      runSpacing: _spacing,
-      alignment: WrapAlignment.end,
-      crossAxisAlignment: WrapCrossAlignment.center,
-      children: [for (final action in actions) _buildAction(action)],
-    );
-  }
-
-  Widget _buildAction(UnitsEntityAction action) {
-    final bool enabled = action.onPressed != null;
-    if (action.iconOnly) {
-      return CtCircularLocateButton(
-        onPressed: action.onPressed,
-        icon: action.icon,
-        tooltip: action.tooltip,
-        semanticLabel: action.label,
-        enabled: enabled,
-      );
-    }
-    if (action.variant == UnitsEntityActionVariant.danger) {
-      return CtDangerTextButton(
-        onPressed: action.onPressed,
-        label: action.label,
-        icon: action.icon,
-        tooltip: action.tooltip,
-        enabled: enabled,
-      );
-    }
-    return CtActionTextButton(
-      onPressed: action.onPressed,
-      label: action.label,
-      icon: action.icon,
-      tooltip: action.tooltip,
-      enabled: enabled,
-    );
-  }
-}
-
-/// Dense chip matching train-dialog resource chips. SPEC/ui/civilian-units-panel.md.
-class _AssignedCostChip extends StatelessWidget {
-  const _AssignedCostChip({required this.child});
-
-  final Widget child;
-
-  @override
-  Widget build(BuildContext context) {
-    return TrainDialogResourceChip(child: child);
   }
 }
